[bug #165645049] this commit arranges users' files
